Working Bounce and stick mechanic
Changes to note:

- added a wall, added tags (leftWall and rightWall)
- added script Bounce Or Stick
- adjusted the preexisting polygon collider, added a circle collider to scissors

Description of added script:

The Bounce or Stick script should make the scissors bounce and spin when the body hits the wall but stick when the tip of the x (circle collider) is triggered.  The public variables BounceForce, BounceGravity, and BounceRotationSpeed can be adjusted for better interaction when more walls are added and the level is expanded upwards.

Current recommended values: BounceForce = 10, BounceRotation = 10, BounceGravity = 1.